ultimate/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.0
crispy-bootstrap5==0.7
Django==4.1.7
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap-v5==1.0.11
django-crispy-forms==2.0
django-jquery==3.1.0
django-mathfilters==1.0.0
django-templatetags==1.1
Pillow==9.4.0
pydantic==1.10.7
soupsieve==2.4
sqlparse==0.4.3
typing_extensions==4.5.0
